Updated Repository with latest code.

+++ Classes/Player.h
@@ -8,7 +8,7 @@
 class Player : public cocos2d::CCSprite{
 public:
 	CC_SYNTHESIZE(int, _curLives, CurLives);
-	CC_SYNTHESIZE(int, _curDamage, CurDamage);
+	CC_SYNTHESIZE(float, _curDamage, CurDamage);
 	CC_SYNTHESIZE(int, _curWeapon, CurWeapon);
 	CC_SYNTHESIZE(float, _attackSpeed, AttackSpeed);
 	CC_SYNTHESIZE(float, _scoreMultiplier, ScoreMultiplier);
